Implemented embroidery items
