+++ src/constants/ActionTypes.ts
@@ -23,3 +23,6 @@ export const SET_LIST_LININGS_MATERIAL = 'setListLiningsMaterial'
 export const SET_STITCH_COLOR = 'setStitchColor'
 export const SET_EMBROIDERY_TYPE_FACE = 'setEmbroideryTypeFace'
 export const SET_EMBROIDERY_POSITION = 'setEmbroideryPosition'
+export const SET_EMBROIDERY_COLOR = 'setEmbroideryColor'
+export const SET_EMBROIDERY_SHADOW_COLOR = 'setEmbroideryShadowColor'
+export const SET_EMBROIDERY_CONTENT = 'setEmbroideryContent'
