+++ src/constants/radioObjs/baseSet.ts
@@ -1,0 +1,87 @@
+export const baseModelObjs = [
+  { label: 'M8型', value: 'm8' },
+  { label: 'M1型', value: 'm1' },
+  { label: 'M2型', value: 'm2' },
+  { label: 'M01型', value: 'm01' },
+]
+
+export const dominantArmObjs = [
+  { label: '右投げ', value: 'rightThrow' },
+  { label: '左投げ', value: 'leftThrow' },
+]
+
+export const mittSizeObjs = [
+  { label: '基本モデルより10mm大きく', value: 'large_5' },
+  { label: '基本モデルより5mm大きく', value: 'large_10' },
+  { label: '基本モデルズバリ', value: 'default' },
+  { label: '基本モデルより5mm小さく', value: 'small_5' },
+  { label: '基本モデルより10mm小さく', value: 'small_10' },
+]
+
+export const mittDepthObjs = [
+  { label: '基本モデルズバリ', value: 'default' },
+  { label: '基本モデルより深く', value: 'deep' },
+  { label: '基本モデルより浅く', value: 'shallow' },
+]
+
+export const backStyleObjs = [
+  { label: 'Aバック', value: 'A_back' },
+  { label: 'Bバック', value: 'B_back' },
+  { label: 'axバック', value: 'ax_back' },
+  { label: 'シェラームーブ', value: 'shellarmove' },
+]
+
+export const padModleObjs = [
+  { label: 'Aバッド', value: 'A_pad' },
+  { label: 'Bバッド', value: 'B_pad' },
+]
+
+export const hardnessObjs = [
+  { label: '硬く', value: 'hard' },
+  { label: '普通', value: 'normal' },
+  { label: '柔らかく', value: 'soft' },
+]
+
+export const thicknessObjs = [
+  { label: '厚く', value: 'thick' },
+  { label: '普通', value: 'normal' },
+  { label: '薄く', value: 'thin' },
+]
+
+export const hamidashiTypeObjs = [
+  { label: '切りハミ(ホワイト)', value: 'kirihami_white' },
+  { label: '切りハミ(ブラック)', value: 'kirihami_black' },
+  { label: 'ハミダシなし', value: 'no_hamidashi' },
+]
+
+export const fingerGuardTypeObjs = [
+  { label: 'スタンダード', value: 'standard_finger_guard' },
+  { label: 'ロング', value: 'long_finger_guard' },
+  { label: '指カバーなし', value: 'no_finger_guard' },
+]
+
+export const zabutonSpongeObjs = [
+  { label: '有り', value: 'zabuton' },
+  { label: '無し', value: 'no_zabuton' },
+]
+
+export const exFunctionObjs = [
+  { label: '有り', value: 'ex_func' },
+  { label: '無し', value: 'no_ex_func' },
+]
+
+export const pinkiePatternObjs = [
+  { label: '有り', value: 'pinkie_pattern' },
+  { label: '無し', value: 'no_pinkie_pattern' },
+]
+
+export const tbEngravedObjs = [
+  { label: '有り', value: 'tb_engraved' },
+  { label: '無し', value: 'no_tb_engraved' },
+]
+
+export const listLiningsMaterialObjs = [
+  { label: 'ムートン(ホワイト)', value: 'mouton_white' },
+  { label: 'ムートン(ブラック)', value: 'mouton_black' },
+  { label: 'ウルトラスエード(パッド)', value: 'ultra_suède' },
+]
