+++ src/constants/radioObjs/coloring.ts
@@ -1,0 +1,53 @@
+export const leatherStringColorObjs = [
+  { label: 'エコロジーブラウン', value: 'ecology_brown', color: '#733b03' },
+  { label: 'ブラウン', value: 'brown', color: '#CC6600' },
+  { label: 'オレンジ', value: 'orange', color: '#ff8c00' },
+  { label: 'ブラック', value: 'black', color: '#000' },
+  { label: 'Jイエロー', value: 'j_yellow', color: '#ffd700' },
+  { label: 'Vオレンジ', value: 'v_orange', color: '#ff4500' },
+  { label: 'Jブルー', value: 'j_blue', color: '#0000DD' },
+  { label: 'レッド', value: 'red', color: '#EE0000' },
+  { label: 'ホワイト', value: 'white', color: '#fff' },
+]
+
+export const leatherColorObjs = [...leatherStringColorObjs, { label: 'Vブラック', value: 'v_black', color: '#525151' }]
+
+export const liningsTypeObjs = [
+  { label: '共牛革(エコロジーブラウン)', value: 'cow_ecology_brown', color: '#733b03' },
+  { label: '共牛革(ブラウン)', value: 'cow_brown', color: '#CC6600' },
+  { label: '共牛革(オレンジ)', value: 'cow_orange', color: '#ff8c00' },
+  { label: '共牛革(ブラック)', value: 'cow_black', color: '#000' },
+  { label: '共牛革(Jイエロー)', value: 'cow_j_yellow', color: '#ffd700' },
+  { label: '共牛革(Vオレンジ)', value: 'cow_v_orange', color: '#ff4500' },
+  { label: '共牛革(Jブルー)', value: 'cow_j_blue', color: '#0000DD' },
+  { label: '共牛革(レッド)', value: 'cow_red', color: '#EE0000' },
+  { label: '共牛革(ホワイト)', value: 'cow_white', color: '#fff' },
+  { label: '共牛革(Vブラック)', value: 'cow_v_black', color: '#525151' },
+  { label: '和牛革(ブラック)', value: 'wagyu_black', color: '#000' },
+  { label: '鹿革', value: 'deer', color: '#000' },
+  { label: 'ウルトラスエード(ブラック)', value: 'ultra_suede', color: '#000' },
+]
+
+export const stitchColorObjs = [
+  { label: 'イエロー', value: 'j_yellow', color: '#ffd700' },
+  { label: 'ブラック', value: 'black', color: '#000' },
+  { label: 'ブルー', value: 'j_blue', color: '#0000DD' },
+  { label: 'レッド', value: 'red', color: '#EE0000' },
+  { label: 'ホワイト', value: 'white', color: '#fff' },
+  { label: 'ブラウン', value: 'brown', color: '#CC6600' },
+  { label: 'オレンジ', value: 'orange', color: '#ff8c00' },
+]
+
+export const hatakeyamaLabelObjs = [
+  { label: '直刺繍C1', value: 'straight_embroidery' },
+  { label: '刺繍ワッペンW1', value: 'emblem_1' },
+  { label: '刺繍ワッペンW2', value: 'emblem_2' },
+  { label: '刺繍ワッペンW3', value: 'emblem_3' },
+  { label: '刺繍ワッペンW4', value: 'emblem_4' },
+  { label: 'シリコンS1', value: 'silicon_1' },
+  { label: 'シリコンS2', value: 'silicon_2' },
+  { label: 'シリコンS3', value: 'silicon_3' },
+  { label: 'シリコンS4', value: 'silicon_4' },
+  { label: 'シリコンS5', value: 'silicon_5' },
+  { label: 'シリコンS6', value: 'silicon_6' },
+]
