+++ src/constants/radioObjs/embroidery.ts
@@ -1,0 +1,36 @@
+export const embroideryTypeFaceObjs = [
+  { label: '楷書体', value: 'regular' },
+  { label: '行書体', value: 'semi_cursive' },
+  { label: 'アルファベット活字体', value: 'alphabet_regular' },
+  { label: 'アルファベット筆記体', value: 'alphabet_cursive' },
+  { label: 'ハタケヤマオリジナル書体', value: 'original' },
+  { label: 'オールドイングリッシュ書体', value: 'old_english' },
+  { label: '刻印', value: 'engraved' },
+]
+
+export const embroideryPositionObjs = [
+  { label: '親指部分', value: 'thumb_finger' },
+  { label: '小指部分', value: 'child_finger' },
+  { label: '甲部分(シェラームーブ)', value: 'shellarmove' },
+  { label: '裏革部分(窓見部分)', value: 'leather_lining_mdd' },
+  { label: '裏革部分(手口部分)', value: 'leather_lining_enter' },
+  { label: 'グラブ袋', value: 'glove_bag' },
+]
+
+export const embroideryColorObjs = [
+  { label: 'ブラック', value: 'black', color: '#000' },
+  { label: 'Vブラック', value: 'v_black', color: '#525151' },
+  { label: 'ネイビー', value: 'navy', color: '#202f55' },
+  { label: 'パープル', value: 'purple', color: '#8123b0' },
+  { label: 'ブルー', value: 'Blue', color: '#0000DD' },
+  { label: 'レッド', value: 'red', color: '#EE0000' },
+]
+
+export const embroideryShadowColorObjs = [
+  { label: 'ブラック', value: 'black', color: '#000' },
+  { label: 'Vブラック', value: 'v_black', color: '#525151' },
+  { label: 'ネイビー', value: 'navy', color: '#202f55' },
+  { label: 'パープル', value: 'purple', color: '#8123b0' },
+  { label: 'ブルー', value: 'Blue', color: '#0000DD' },
+  { label: 'レッド', value: 'red', color: '#EE0000' },
+]
